+++ fourtoon-cookie/src/pages/DiaryTimelinePage/DiaryComponent/DiaryComponent.tsx
@@ -1,13 +1,12 @@
 import React from "react";
-import { View } from "react-native";
+import { StyleSheet, View } from "react-native";
 import { Diary, DiaryStatus } from "../../../types/diary";
-import DiaryContentsLayout from "./DiaryContentsLayout/DiaryContentsLayout";
-import DiaryPaintingImagesLayout from "./DiaryPaintingImagesLayout/DiaryPaintingImagesLayout";
-import * as S from './DiaryComponent.styled';
-import Footer from "./Footer/Footer";
-import Header from "./Header/Header";
-import DiaryPaintingImageLoadingLayout from "./DiaryPaintingImageLoadingLayout/DiaryPaintingImageLoadingLayout";
-import DiaryPaintingImageFailedLayout from "./DiaryPaintingImageFailedLayout/DiaryPaintingImageFailedLayout";
+import DiaryContentsLayout from "./DiaryContentsLayout";
+import DiaryPaintingImagesLayout from "./DiaryPaintingImagesLayout";
+import Footer from "./Footer";
+import Header from "./Header";
+import DiaryPaintingImageLoadingLayout from "./DiaryPaintingImageLoadingLayout";
+import DiaryPaintingImageFailedLayout from "./DiaryPaintingImageFailedLayout";
 import { useDiaryById } from "../../../hooks/server/diary";
 
 export interface DiaryProps {
@@ -23,7 +22,7 @@ const DiaryComponent = (props: DiaryProps) => {
     const { content, isFavorite, diaryDate, paintingImageUrls, characterId, diaryStatus } = diary;
 
     return (
-        <View style={S.styles.container}>
+        <View style={styles.container}>
             <Header
                 diaryId={diaryId}
                 characterId={diary.characterId}
@@ -78,3 +77,74 @@ const DiaryBody = (props: DiaryBodyProps) => {
         <DiaryPaintingImageFailedLayout />
     );
 }
+
+const styles = StyleSheet.create({
+    container: {
+        display: 'flex',
+        flexDirection: 'column',
+        alignItems: 'flex-start',
+        padding: 16,
+        marginLeft: 10,
+        marginRight: 10,
+        marginBottom: 10,
+        gap: 16,
+        width: 'auto',
+        height: 'auto',
+        backgroundColor: '#FFFFFF',
+        borderRadius: 30,
+    },
+    contents: {
+        flexDirection: 'column',
+        marginBottom: 10,
+    },
+    footer: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        marginBottom: 10,
+    },
+    likeImage: {
+        width: 30,
+        height: 30,
+    },
+    footerLikeButton: {
+        flex: 1,
+    },
+    settingButtons: {
+        flex: 3,
+        flexDirection: 'row',
+        justifyContent: 'flex-end',
+    },
+    button: {
+        padding: 5,
+        borderRadius: 5,
+        borderWidth: 1,
+        borderColor: 'gray',
+        backgroundColor: 'gray',
+        textAlign: 'center',
+        marginLeft: 5,
+    },
+    modalContainer: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+        backgroundColor: 'rgba(0, 0, 0, 0.5)',
+    },
+    modalContent: {
+        width: 300,
+        padding: 20,
+        backgroundColor: 'white',
+        borderRadius: 10,
+        alignItems: 'center',
+    },
+    modalText: {
+        marginBottom: 20,
+        fontSize: 18,
+        textAlign: 'center',
+    },
+    modalButtons: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        width: '100%',
+    },
+});
