--- fourtoon-cookie/src/pages/DiaryTimelinePage/DiaryEmpty/DiaryEmpty.styled.ts
@@ -1,21 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-    },
-    text: {
-        fontSize: 24,
-        marginLeft: 10,
-        color: '#000'
-    },
-    button: {
-        flexDirection: 'row',
-        justifyContent: 'center',
-    },
-    image: {
-        width: 20,
-        height: 20
-    }
-});
