--- fourtoon-cookie/src/pages/DiaryTimelinePage/DiaryTimelinePage.styled.ts
@@ -1,4 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-});
