+++ fourtoon-cookie/src/pages/DiaryTimelinePage/DiaryTimelinePage.tsx
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from "react";
 import {FlatList} from "react-native";
-import Header from "./Header/Header";
-import DiaryEmpty from "./DiaryEmpty/DiaryEmpty";
+import Header from "./Header";
+import DiaryEmpty from "./DiaryEmpty";
 import DiaryComponent from "./DiaryComponent/DiaryComponent";
 import MainPageLayout from "../../components/layout/MainPageLayout/MainPageLayout";
 import { FOOTER_STATE } from "../../components/layout/MainPageLayout/Footer/Footer";
